renamed the feature-details component to feature-panel-header
