+++ web-ng/src/app/components/feature-panel-header/feature-panel-header.component.scss
