+++ web-ng/src/app/components/feature-panel/feature-panel.component.html
@@ -18,10 +18,10 @@ limitations under the License.
     TODO: create separate component for layer as a header instead of 
     using LayerListItem here.
   -->
-  <ground-feature-details
+  <ground-feature-panel-header
     [layer]="layer"
     [actionsType]="featureHeaderActionType.BACK"
-  ></ground-feature-details>
+  ></ground-feature-panel-header>
 </mat-list>
 <div *ngIf="layer$ | async as layer">
   <button
